simplify the MCP test case
