+++ intelli/test/integration/mcp_math_server.py
@@ -1,0 +1,25 @@
+from mcp.server.fastmcp import FastMCP
+
+# Create an MCP server
+mcp = FastMCP("MathTools")
+
+# Add an addition tool
+@mcp.tool()
+def add(a: int, b: int) -> int:
+    """Add two numbers"""
+    return a + b
+
+# Add a subtraction tool
+@mcp.tool()
+def subtract(a: int, b: int) -> int:
+    """Subtract two numbers"""
+    return a - b
+
+# Add a multiplication tool
+@mcp.tool()
+def multiply(a: int, b: int) -> int:
+    """Multiply two numbers"""
+    return a * b
+
+if __name__ == "__main__":
+    mcp.run(transport="stdio")
