+++ intelli/test/integration/test_mcp_openai_flow.py
@@ -1,6 +1,5 @@
 import unittest
 import os
-import tempfile
 import sys
 import json
 import asyncio
@@ -15,43 +14,9 @@ load_dotenv()
 
 class TestMCPOpenAIFlow(unittest.TestCase):
     def setUp(self):
-        # Create a temporary directory for our MCP server script
-        self.temp_dir = tempfile.TemporaryDirectory()
-        
-        # Create the MCP server script similar to the one provided in the user request
-        self.server_path = os.path.join(self.temp_dir.name, "mcp_server.py")
-        with open(self.server_path, "w") as f:
-            f.write("""
-from mcp.server.fastmcp import FastMCP
-
-# Create an MCP server
-mcp = FastMCP("Demo")
-
-# Add an addition tool
-@mcp.tool()
-def add(a: int, b: int) -> int:
-    \"\"\"Add two numbers\"\"\"
-    return a + b
-
-# Add a subtraction tool
-@mcp.tool()
-def subtract(a: int, b: int) -> int:
-    \"\"\"Subtract two numbers\"\"\"
-    return a - b
-
-# Add a multiplication tool
-@mcp.tool()
-def multiply(a: int, b: int) -> int:
-    \"\"\"Multiply two numbers\"\"\"
-    return a * b
-
-if __name__ == "__main__":
-    mcp.run(transport="stdio")
-""")
-    
-    def tearDown(self):
-        # Remove temporary directory and its contents
-        self.temp_dir.cleanup()
+        # Get the path to the MCP server file
+        current_dir = os.path.dirname(os.path.abspath(__file__))
+        self.server_path = os.path.join(current_dir, "mcp_math_server.py")
     
     def test_openai_mcp_flow(self):
         """Test a flow that combines OpenAI with MCP for arithmetic operations"""
